fix(issues) Fix duplicate requests in org issues
Changing a search term shouldn't re-fetch the processing issues unless
the projects involved have changed. I've also removed a bunch of aborted
requests when changing between saved searches. I've not been able to
remove them all though.

Currently when moving between saved searches that have different
projects there is one aborted request which comes from the selection
projects being updated after the savedSearch state is updated. I tried
a bunch of different ways to solve this but each solution that resulted
in 0 aborted requests broke a different use case.

I've moved the saved search selection to be an action instead of
a direct link as this helps cut down on the aborted requests and allows
each issue list to generate the appropriate URLs without having to embed
that logic in the SavedSearchSelector component.

Refs APP-1028