api(feat): Modify OrganizationEnvironmentEndpoint to filter by projects and respect visibility (APP-1035)
This endpoint was returning all environments for an `Organization`, which isn't great for orgs with
many environments, and projects that are only associated with a small number of environment. It now
accepts our standard list of project ids and filters based on those, or if not passed then all
projects that the user has access to.

Also added the `visibility` querystring param so that we can filter by visibility in the same way
as the `ProjectEnvironment` endpoint. An environment will be visible if it's not hidden in at least
one of the related projects.